USB: f_rmnet_smd: Process rmnet_disable on usb disconnect, if smd_open fail
If smd_open fails during USB connect and when "adb root" command is
executed, USB gets disconnected, and since the online flag is 0,
the disconnect process will not be carried out and the platform driver
is not unregistered. On USB connect, as the the_pdrv_used flag is reset
to 0, we endup in registering the platform_driver multiple times because
of which crashes are observed.

Hence, in rmnet_disable() function, don't check for the online flag, so
that the disconnect work will get processed during USB disconnect and
platform driver unregister is carried out.

CRs-Fixed: 414886
Signed-off-by: Rajkumar Raghupathy <raghup@codeaurora.org>
(cherry picked from commit 28930dc7bebbec841fbff380082067d596f15b22)

Change-Id: I5f616567120ed73c7f7c1f961b994ce6b2f00ffd
Signed-off-by: Gopal Goberu <ggober@codeaurora.org>