USB: android: Handle error conditions if usb_add_config fails
When bind_config() of any function driver fails, we go ahead and do a
usb_gadget_connect() due to which remove_config is called during
next disconnect process and a crash is observed as we delete the
config->list once again.

So if the bind_config of any function driver fails, don't call
usb_gadget_connect(), instead handle the possible error conditions and
notify disconnect event to userspace so that the composition switches
back to default composition.

CRs-Fixed: 418180
Change-Id: If1876a32612074ac69f6cc134be0c93fbe781452
Signed-off-by: Rajkumar Raghupathy <raghup@codeaurora.org>