msm: cpr: Update threshold as per voltage step size
Down threshold decides the number of error steps which
are needed for CPR to raise an interrupt. It uses the
equation: no interrupt if error_steps < up/dn threshold.
When using external buck regulator the step size is half
of that on PMIC based regulator. So, adjust the down
threshold accordingly to keep the same effective
threshold in terms of voltage.

Change-Id: Iaef9af404d7f629f40b4a07e03e1d8cff67bb756
Signed-off-by: Kaushal Kumar <kaushalk@codeaurora.org>