msm: cpr: Use preset value of Turbo Vmin in 8x25Q
The floor fuse is used in 8x25(v2) to fix the L2 cache
Vmin marginality. In 8x25Q, the Vmin for both nominal
and turbo is fixed so do not depend on floor fuse since
these fused bits could possibly get re-used for some
other purpose leading to error in turbo Vmin setting.

Change-Id: I2202d23a5b552434082c406bcafd0f2bc4b0a6fe
Signed-off-by: Kaushal Kumar <kaushalk@codeaurora.org>