regulator: ncp6335d: API to configure the regulator during reboot
In cases when the device is not completely turned off (reboot)
the regulator registers are not reset. This can cause the device to
boot at the same voltage at which it went down, leading to
instability. Add an API to set the default voltage.

Also, call this API before the APPS goes for a restart on 8x25q.

CRs-Fixed: 444285
Change-Id: If023422e32acfe20d8cc0eafa626ad7331628ab2
Signed-off-by: Anirudh Ghayal <aghayal@codeaurora.org>