msm: kgsl: Don't change the sglen of the memdesc if allocation fails
If the page_alloc loop fails, we changed the length of the scatterlist
for the memdesc to the last good page so we could have an efficient
loop through kgsl_sharedmem_free.  Unfortunately, we were also using
the sglen to decide between using kmem or vmalloc memory for the sg
list in the first place. There is a chance that one could start out
with a large sg list from vmalloc, but if it failed early, then we
could truncate the size and attempt to kfree it.

Zero the stlist at init time and use a length == 0 as an indicator
that the list might be truncated.

Change-Id: Ic0dedbadc0007b5e3c338151517c95999ed523a3
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>
Signed-off-by: Vinay Roy <vroy@codeaurora.org>