regulator: onsemi: Register for syscore ops
Register the shutdown hook in syscore_ops to make sure that
the buck goes down with the default voltage. This is required
in the case of a restart where the input voltage to the buck
is not turned off. It is not possible to do this operation in
the restart handler (msm_pm_restart) as preemption is disabled
at that point.

Also make sure that the voltage cannot be set after the
shutdown hook is executed.

CRs-Fixed: 458339
Change-Id: I01f0338dfc9ad1c6c855f14b7684fae355c72257
Signed-off-by: Anirudh Ghayal <aghayal@codeaurora.org>