msm: kgsl: Start Busy Timer before resetting the busy counter
When GPU is 100% busy, starting the busy counter before the begining
of next busy start makes the bus interval exceed total time. Another
possibility is that we started the busy_start_time for the next sample
but since we reset the perfmon control rtegister after sometime, when
we stop the busy_start_time for the current submission, the GPU busy
ticks is still. These two conditions could result in negative idle
stats and cause Busy timer expiry.

The change ensures that we start busy timer first and then reset the
perfmon control registers, so that total time is not less than busy
interval. Also add some delta of 100 microsec to complensate for the
delay in starting the two clocks.

Change-Id: Id7f4cb20916b39740ae88849bf56857b821ec12d
CRs-fixed: 460011
Signed-off-by: Vinay Roy <vroy@codeaurora.org>